Update docs, move subscribeAll to different file
